AP-5613: REST Client nodes (Labs) - follow-up project of famous KREST community nodes
Task-Url: https://knime-com.atlassian.net/browse/AP-5613

Changed authentication to be applied after the headers set.
Improved the dialog in regard of the table editing (now the request table keys follow the selection of the template).
Known problem (with incomplete fix): when a dialog opened with no rows, adding two rows and starting editing one of them and clicking to the other row will set the content to the other row instead of keeping both empty or leaving the content in the original row.