AP-20595: Ensure that a CXF Bus is initialized before creating the CXF Client
When using NTLM authentication, the Bus is closed after the Node
execution, which results in execution failures in subsequent runs, as
creating the Client requires a Bus, and the Bus created in the Activator
is no longer available.
By setting up the Bus before creating the client, we avoid this issue.